refactor(theme): added onGlobalSetup to call composition on Setup
